+++ CHANGES.md
@@ -1,3 +1,8 @@
+##
+
+- Windows bundle
+
+
 ## 0.0.2
 
 - Fix infinite nesting of directories during GEM installation